requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
fastapi==0.118.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
pillow==11.3.0
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
